requirements.txt
google.generativeai
langchain-google-genai
langchain
langchain_community
jupyter
python-dotenv
streamlit
textblob
time
nltk